Split the forex module into smaller modules
