+++ src/xrc/src/forex.rs
@@ -1,14 +1,19 @@
 mod australia;
+mod bosnia_herzegovina;
+mod canada;
+mod europe;
+mod myanmar;
+mod singapore;
+mod uzbekistan;
 
-use chrono::naive::NaiveDateTime;
 use ic_cdk::export::candid::{
     decode_args, decode_one, encode_args, encode_one, CandidType, Deserialize, Error as CandidError,
 };
 use ic_xrc_types::{Asset, AssetClass, ExchangeRateError};
 use std::cmp::min;
+use std::collections::HashMap;
 use std::collections::{HashSet, VecDeque};
 use std::mem::size_of_val;
-use std::{collections::HashMap, convert::TryInto};
 
 use crate::{
     median, standard_deviation, utils, AllocatedBytes, ExtractError, QueriedExchangeRate, ONE_KIB,
@@ -753,635 +758,14 @@ trait IsForex {
     }
 }
 
-/// Monetary Authority Of Singapore
-#[derive(Deserialize)]
-struct MonetaryAuthorityOfSingaporeResponse {
-    result: MonetaryAuthorityOfSingaporeResponseResult,
-}
-
-#[derive(Deserialize)]
-struct MonetaryAuthorityOfSingaporeResponseResult {
-    records: Vec<HashMap<String, String>>,
-}
-
-impl IsForex for MonetaryAuthorityOfSingapore {
-    fn format_timestamp(&self, timestamp: u64) -> String {
-        format!(
-            "{}",
-            NaiveDateTime::from_timestamp(timestamp.try_into().unwrap_or(0), 0).format("%Y-%m-%d")
-        )
-    }
-
-    fn extract_rate(&self, bytes: &[u8], timestamp: u64) -> Result<ForexRateMap, ExtractError> {
-        let response = serde_json::from_slice::<MonetaryAuthorityOfSingaporeResponse>(bytes)
-            .map_err(|err| ExtractError::json_deserialize(bytes, err.to_string()))?;
-        let timestamp = (timestamp / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-
-        let map = response.result.records.get(0).ok_or_else(|| {
-            ExtractError::json_deserialize(bytes, "Missing record index".to_string())
-        })?;
-
-        let extracted_timestamp = {
-            let maybe_end_of_day = map.get("end_of_day");
-            let end_of_day = match maybe_end_of_day {
-                Some(end_of_day) => NaiveDateTime::parse_from_str(
-                    &(end_of_day.to_string() + " 00:00:00"),
-                    "%Y-%m-%d %H:%M:%S",
-                )
-                .unwrap_or_else(|_| NaiveDateTime::from_timestamp(0, 0)),
-                None => NaiveDateTime::from_timestamp(0, 0),
-            };
-            end_of_day.timestamp() as u64
-        };
-
-        if extracted_timestamp != timestamp {
-            return Err(ExtractError::RateNotFound {
-                filter: "Invalid timestamp".to_string(),
-            });
-        }
-
-        let mut values = map
-            .iter()
-            .filter_map(|(key, value)| {
-                if !key.contains("_sgd") {
-                    return None;
-                }
-
-                match value.parse::<f64>() {
-                    Ok(rate) => match key.split('_').next() {
-                        Some(symbol) => {
-                            let scaled_rate = if key.ends_with("_100") {
-                                (rate * (RATE_UNIT as f64 / 100.0)) as u64
-                            } else {
-                                (rate * RATE_UNIT as f64) as u64
-                            };
-
-                            Some((symbol.to_uppercase(), scaled_rate))
-                        }
-                        None => None,
-                    },
-                    Err(_) => None,
-                }
-            })
-            .collect::<ForexRateMap>();
-        values.insert("SGD".to_string(), RATE_UNIT);
-        self.normalize_to_usd(&values)
-    }
-
-    fn get_base_url(&self) -> &str {
-        "https://eservices.mas.gov.sg/api/action/datastore/search.json?resource_id=95932927-c8bc-4e7a-b484-68a66a24edfe&limit=100&filters[end_of_day]=DATE"
-    }
-
-    fn supports_ipv6(&self) -> bool {
-        true
-    }
-
-    fn get_utc_offset(&self) -> i16 {
-        8
-    }
-
-    fn max_response_bytes(&self) -> u64 {
-        // 3 KiB
-        ONE_KIB * 3
-    }
-}
-
-/// Central Bank of Myanmar
-#[derive(Debug, Deserialize)]
-struct CentralBankOfMyanmarResponse {
-    timestamp: u64,
-    rates: HashMap<String, String>,
-}
-
-impl IsForex for CentralBankOfMyanmar {
-    fn format_timestamp(&self, timestamp: u64) -> String {
-        format!(
-            "{}",
-            NaiveDateTime::from_timestamp(timestamp.try_into().unwrap_or(0), 0).format("%d-%m-%Y")
-        )
-    }
-
-    fn extract_rate(&self, bytes: &[u8], timestamp: u64) -> Result<ForexRateMap, ExtractError> {
-        let response = serde_json::from_slice::<CentralBankOfMyanmarResponse>(bytes)
-            .map_err(|err| ExtractError::json_deserialize(bytes, err.to_string()))?;
-        let timestamp = (timestamp / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-
-        if response.timestamp != timestamp {
-            return Err(ExtractError::RateNotFound {
-                filter: "Invalid timestamp".to_string(),
-            });
-        }
-
-        let values = response
-            .rates
-            .iter()
-            .filter_map(|(asset, rate)| {
-                let parsed = rate.replace(',', "").parse::<f64>().ok()?;
-                let mut rate = (parsed * RATE_UNIT as f64) as u64;
-                if asset == "JPY" {
-                    rate /= 100;
-                }
-                Some((asset.to_uppercase(), rate))
-            })
-            .collect::<ForexRateMap>();
-        self.normalize_to_usd(&values)
-    }
-
-    fn get_base_url(&self) -> &str {
-        "https://forex.cbm.gov.mm/api/history/DATE"
-    }
-
-    fn supports_ipv6(&self) -> bool {
-        true
-    }
-
-    fn get_utc_offset(&self) -> i16 {
-        // Myanmar timezone is UTC+6.5. To avoid using floating point types here, we use a truncated offset.
-        6
-    }
-
-    fn max_response_bytes(&self) -> u64 {
-        // 3KiB - this is need to get past the http body size limit
-        ONE_KIB * 3
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct CentralBankOfBosniaHerzegovinaResponseCurrencyExchangeItem {
-    #[serde(rename(deserialize = "AlphaCode"))]
-    alpha_code: String,
-    #[serde(rename(deserialize = "Units"))]
-    units: String,
-    #[serde(rename(deserialize = "Middle"))]
-    middle: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct CentralBankOfBosniaHerzegovinaResponse {
-    #[serde(rename(deserialize = "CurrencyExchangeItems"))]
-    currency_exchange_items: Vec<CentralBankOfBosniaHerzegovinaResponseCurrencyExchangeItem>,
-    #[serde(rename(deserialize = "Date"))]
-    date: String,
-}
-
-/// Central Bank of Bosnia-Herzegovina
-impl IsForex for CentralBankOfBosniaHerzegovina {
-    fn format_timestamp(&self, timestamp: u64) -> String {
-        format!(
-            "{}",
-            NaiveDateTime::from_timestamp(timestamp.try_into().unwrap_or(0), 0).format("%m-%d-%Y")
-        )
-    }
-
-    fn extract_rate(&self, bytes: &[u8], timestamp: u64) -> Result<ForexRateMap, ExtractError> {
-        let response = serde_json::from_slice::<CentralBankOfBosniaHerzegovinaResponse>(bytes)
-            .map_err(|err| ExtractError::json_deserialize(bytes, err.to_string()))?;
-        let timestamp = (timestamp / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-
-        let extracted_timestamp = NaiveDateTime::parse_from_str(&response.date, "%Y-%m-%dT%H:%M:%S")
-            .unwrap_or_else(|_| NaiveDateTime::from_timestamp(0, 0))
-            .timestamp() as u64;
-        if extracted_timestamp != timestamp {
-            return Err(ExtractError::RateNotFound {
-                filter: "Invalid timestamp".to_string(),
-            });
-        }
-
-        let values = response
-            .currency_exchange_items
-            .iter()
-            .filter_map(|item| {
-                let units = item.units.parse::<u64>().ok()?;
-                let middle = item.middle.replace(',', ".").parse::<f64>().ok()?;
-                let rate = ((middle * RATE_UNIT as f64) / units as f64) as u64;
-
-                Some((item.alpha_code.clone(), rate))
-            })
-            .collect::<ForexRateMap>();
-        self.normalize_to_usd(&values)
-    }
-
-    fn get_base_url(&self) -> &str {
-        "https://www.cbbh.ba/CurrencyExchange/GetJson?date=DATE%2000%3A00%3A00"
-    }
-
-    fn supports_ipv6(&self) -> bool {
-        true
-    }
-
-    fn get_utc_offset(&self) -> i16 {
-        1
-    }
-
-    fn offset_timestamp_for_query(&self, timestamp: u64) -> u64 {
-        // To fetch the rates for day X, Central Bank of Bosnia-Herzgovina expects the supplied argument to be the day of X+1.
-        ((timestamp / SECONDS_PER_DAY) + 1) * SECONDS_PER_DAY
-    }
-
-    /// Responses are between 20-25 KiB. Set to 30 to give some leeway.
-    fn max_response_bytes(&self) -> u64 {
-        // 30 KiB
-        ONE_KIB * 30
-    }
-}
-
-#[derive(Deserialize, Debug)]
-enum XmlEcbOptions {
-    #[serde(rename = "subject")]
-    Subject(String),
-    #[serde(rename = "Sender")]
-    Sender(XmlEcbSender),
-    Cube(XmlEcbOuterCube),
-}
-
-#[derive(Deserialize, Debug)]
-struct XmlEcbSender {
-    #[serde(rename = "name")]
-    _name: String,
-}
-
-#[derive(Deserialize, Debug)]
-struct XmlEcbCubeObject {
-    time: String,
-    #[serde(rename = "$value")]
-    cubes: Vec<XmlEcbCube>,
-}
-
-#[derive(Deserialize, Debug)]
-struct XmlEcbCube {
-    currency: String,
-    rate: f64,
-}
-
-#[derive(Deserialize, Debug)]
-struct XmlEcbOuterCube {
-    #[serde(rename = "Cube")]
-    cube: XmlEcbCubeObject,
-}
-
-#[derive(Deserialize, Debug)]
-struct XmlEcbEnvelope {
-    #[serde(rename = "$value")]
-    entries: Vec<XmlEcbOptions>,
-}
-
-/// European Central Bank
-impl IsForex for EuropeanCentralBank {
-    fn format_timestamp(&self, _timestamp: u64) -> String {
-        // ECB does not take a timestamp/date as an argument. It always returns the latest date.
-        "".to_string()
-    }
-
-    fn extract_rate(&self, bytes: &[u8], timestamp: u64) -> Result<ForexRateMap, ExtractError> {
-        let timestamp = (timestamp / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-
-        let data: XmlEcbEnvelope = serde_xml_rs::from_reader(bytes)
-            .map_err(|e| ExtractError::XmlDeserialize(format!("{:?}", e)))?;
-
-        if let XmlEcbOptions::Cube(cubes) = data
-            .entries
-            .iter()
-            .find(|e| matches!(e, XmlEcbOptions::Cube(_)))
-            .unwrap_or(&XmlEcbOptions::Cube(XmlEcbOuterCube {
-                cube: XmlEcbCubeObject {
-                    time: "0".to_string(),
-                    cubes: vec![],
-                },
-            }))
-        {
-            let extracted_timestamp = NaiveDateTime::parse_from_str(
-                &(cubes.cube.time.clone() + " 00:00:00"),
-                "%Y-%m-%d %H:%M:%S",
-            )
-            .unwrap_or_else(|_| NaiveDateTime::from_timestamp(0, 0))
-            .timestamp() as u64;
-
-            if extracted_timestamp != timestamp {
-                Err(ExtractError::RateNotFound {
-                    filter: "Invalid timestamp".to_string(),
-                })
-            } else {
-                let mut values: ForexRateMap = cubes
-                    .cube
-                    .cubes
-                    .iter()
-                    .map(|cube| {
-                        (
-                            cube.currency.to_uppercase(),
-                            ((1.0 / cube.rate) * RATE_UNIT as f64) as u64,
-                        )
-                    })
-                    .collect();
-                values.insert("EUR".to_string(), RATE_UNIT);
-                self.normalize_to_usd(&values)
-            }
-        } else {
-            Err(ExtractError::XmlDeserialize(
-                String::from_utf8(bytes.to_vec()).unwrap_or_default(),
-            ))
-        }
-    }
-
-    fn get_base_url(&self) -> &str {
-        "https://www.ecb.europa.eu/stats/eurofxref/eurofxref-daily.xml"
-    }
-
-    fn get_utc_offset(&self) -> i16 {
-        1
-    }
-
-    fn max_response_bytes(&self) -> u64 {
-        ONE_KIB * 3
-    }
-}
-
-/// Bank of Canada
-#[derive(Debug, Deserialize)]
-struct BankOfCanadaResponseSeriesDetail {
-    label: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct BankOfCanadaResponseObservation {
-    d: String,
-    #[serde(flatten)]
-    rates: HashMap<String, BankOfCanadaResponseObservationValue>,
-}
-
-#[derive(Debug, Deserialize)]
-struct BankOfCanadaResponseObservationValue {
-    v: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct BankOfCanadaResponse {
-    #[serde(rename(deserialize = "seriesDetail"))]
-    series_detail: HashMap<String, BankOfCanadaResponseSeriesDetail>,
-    observations: Vec<BankOfCanadaResponseObservation>,
-}
-
-impl IsForex for BankOfCanada {
-    fn format_timestamp(&self, timestamp: u64) -> String {
-        format!(
-            "{}",
-            NaiveDateTime::from_timestamp(timestamp.try_into().unwrap_or(0), 0).format("%Y-%m-%d")
-        )
-    }
-
-    fn extract_rate(&self, bytes: &[u8], timestamp: u64) -> Result<ForexRateMap, ExtractError> {
-        let response = serde_json::from_slice::<BankOfCanadaResponse>(bytes)
-            .map_err(|err| ExtractError::json_deserialize(bytes, err.to_string()))?;
-
-        let timestamp = (timestamp / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-        let mut extracted_timestamp: u64;
-        let mut values = ForexRateMap::new();
-        for observation in response.observations.iter() {
-            extracted_timestamp = NaiveDateTime::parse_from_str(
-                &(observation.d.to_string() + " 00:00:00"),
-                "%Y-%m-%d %H:%M:%S",
-            )
-            .unwrap_or_else(|_| NaiveDateTime::from_timestamp(0, 0))
-            .timestamp() as u64;
-            if extracted_timestamp != timestamp {
-                return Err(ExtractError::RateNotFound {
-                    filter: "Invalid timestamp".to_string(),
-                });
-            }
-
-            observation.rates.iter().for_each(|(key, value)| {
-                let detail = match response.series_detail.get(key) {
-                    Some(detail) => detail,
-                    None => return,
-                };
-
-                let symbol = match detail.label.split('/').next() {
-                    Some(symbol) => symbol,
-                    None => return,
-                };
-
-                let value = match value.v.parse::<f64>() {
-                    Ok(value) => value,
-                    Err(_) => return,
-                };
-
-                values.insert(symbol.to_uppercase(), (value * RATE_UNIT as f64) as u64);
-            });
-        }
-
-        values.insert("CAD".to_string(), RATE_UNIT);
-        self.normalize_to_usd(&values)
-    }
-
-    fn get_base_url(&self) -> &str {
-        "https://www.bankofcanada.ca/valet/observations/group/FX_RATES_DAILY/json?start_date=DATE&end_date=DATE"
-    }
-
-    fn get_utc_offset(&self) -> i16 {
-        // Using the westmost timezone (PST)
-        -8
-    }
-
-    fn max_response_bytes(&self) -> u64 {
-        ONE_KIB * 10
-    }
-}
-
-/// Central Bank of Uzbekistan
-#[derive(Debug, Deserialize)]
-struct CentralBankOfUzbekistanDetail {
-    #[serde(rename(deserialize = "Ccy"))]
-    currency: String,
-    #[serde(rename(deserialize = "Rate"))]
-    rate: String,
-    #[serde(rename(deserialize = "Date"))]
-    date: String,
-}
-
-impl IsForex for CentralBankOfUzbekistan {
-    fn format_timestamp(&self, timestamp: u64) -> String {
-        format!(
-            "{}",
-            NaiveDateTime::from_timestamp(timestamp.try_into().unwrap_or(0), 0).format("%Y-%m-%d")
-        )
-    }
-
-    fn extract_rate(&self, bytes: &[u8], timestamp: u64) -> Result<ForexRateMap, ExtractError> {
-        let response = serde_json::from_slice::<Vec<CentralBankOfUzbekistanDetail>>(bytes)
-            .map_err(|err| ExtractError::json_deserialize(bytes, err.to_string()))?;
-
-        let timestamp = (timestamp / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-        let mut values = ForexRateMap::new();
-
-        for detail in response {
-            let extracted_timestamp =
-                NaiveDateTime::parse_from_str(&(detail.date + " 00:00:00"), "%d.%m.%Y %H:%M:%S")
-                    .unwrap_or_else(|_| NaiveDateTime::from_timestamp(0, 0))
-                    .timestamp() as u64;
-            if extracted_timestamp != timestamp {
-                return Err(ExtractError::RateNotFound {
-                    filter: "Invalid timestamp".to_string(),
-                });
-            }
-
-            let rate = match detail.rate.parse::<f64>() {
-                Ok(rate) => (rate * RATE_UNIT as f64) as u64,
-                Err(_) => continue,
-            };
-
-            values.insert(detail.currency, rate);
-        }
-
-        self.normalize_to_usd(&values)
-    }
-
-    fn get_base_url(&self) -> &str {
-        "https://cbu.uz/ru/arkhiv-kursov-valyut/json/all/DATE/"
-    }
-
-    fn get_utc_offset(&self) -> i16 {
-        5
-    }
-
-    fn max_response_bytes(&self) -> u64 {
-        ONE_KIB * 30
-    }
-}
-
 #[cfg(test)]
 mod test {
     use ic_xrc_types::{ExchangeRate, ExchangeRateMetadata};
     use maplit::hashmap;
 
-    use crate::utils::test::load_file;
     use crate::DECIMALS;
 
     use super::*;
-
-    /// The function test if the macro correctly generates the
-    /// [core::fmt::Display] trait's implementation for [Forex].
-    #[test]
-    fn forex_to_string_returns_name() {
-        let forex = Forex::MonetaryAuthorityOfSingapore(MonetaryAuthorityOfSingapore);
-        assert_eq!(forex.to_string(), "MonetaryAuthorityOfSingapore");
-        let forex = Forex::CentralBankOfMyanmar(CentralBankOfMyanmar);
-        assert_eq!(forex.to_string(), "CentralBankOfMyanmar");
-        let forex = Forex::CentralBankOfBosniaHerzegovina(CentralBankOfBosniaHerzegovina);
-        assert_eq!(forex.to_string(), "CentralBankOfBosniaHerzegovina");
-        let forex = Forex::EuropeanCentralBank(EuropeanCentralBank);
-        assert_eq!(forex.to_string(), "EuropeanCentralBank");
-        let forex = Forex::BankOfCanada(BankOfCanada);
-        assert_eq!(forex.to_string(), "BankOfCanada");
-        let forex = Forex::CentralBankOfUzbekistan(CentralBankOfUzbekistan);
-        assert_eq!(forex.to_string(), "CentralBankOfUzbekistan");
-    }
-
-    /// The function tests if the macro correctly generates derive copies by
-    /// verifying that the forex sources return the correct query string.
-    #[test]
-    fn query_string() {
-        // Note that the hours/minutes/seconds are ignored, setting the considered timestamp to 1661472000.
-        let timestamp = 1661524016;
-        let singapore = MonetaryAuthorityOfSingapore;
-        let query_string = singapore.get_url(timestamp);
-        assert_eq!(query_string, "https://eservices.mas.gov.sg/api/action/datastore/search.json?resource_id=95932927-c8bc-4e7a-b484-68a66a24edfe&limit=100&filters[end_of_day]=2022-08-26");
-        let myanmar = CentralBankOfMyanmar;
-        let query_string = myanmar.get_url(timestamp);
-        assert_eq!(
-            query_string,
-            "https://forex.cbm.gov.mm/api/history/26-08-2022"
-        );
-        let bosnia = CentralBankOfBosniaHerzegovina;
-        let query_string = bosnia.get_url(timestamp);
-        assert_eq!(
-            query_string,
-            "https://www.cbbh.ba/CurrencyExchange/GetJson?date=08-26-2022%2000%3A00%3A00"
-        );
-        let ecb = EuropeanCentralBank;
-        let query_string = ecb.get_url(timestamp);
-        assert_eq!(
-            query_string,
-            "https://www.ecb.europa.eu/stats/eurofxref/eurofxref-daily.xml"
-        );
-        let canada = BankOfCanada;
-        let query_string = canada.get_url(timestamp);
-        assert_eq!(
-            query_string,
-            "https://www.bankofcanada.ca/valet/observations/group/FX_RATES_DAILY/json?start_date=2022-08-26&end_date=2022-08-26"
-        );
-        let uzbekistan = CentralBankOfUzbekistan;
-        let query_string = uzbekistan.get_url(timestamp);
-        assert_eq!(
-            query_string,
-            "https://cbu.uz/ru/arkhiv-kursov-valyut/json/all/2022-08-26/"
-        );
-    }
-
-    /// The function tests if the [MonetaryAuthorityOfSingapore] struct returns the correct forex rate.
-    #[test]
-    fn extract_rate_from_singapore() {
-        let singapore = MonetaryAuthorityOfSingapore;
-        let query_response = load_file("test-data/forex/monetary-authority-of-singapore.json");
-        let timestamp: u64 = 1656374400;
-        let extracted_rates = singapore.extract_rate(&query_response, timestamp);
-
-        assert!(matches!(extracted_rates, Ok(ref rates) if rates["EUR"] == 1_058_173_944));
-        assert!(matches!(extracted_rates, Ok(ref rates) if rates["JPY"] == 7_390_111));
-    }
-
-    /// The function tests if the [CentralBankOfMyanmar] struct returns the correct forex rate.
-    #[test]
-    fn extract_rate_from_myanmar() {
-        let myanmar = CentralBankOfMyanmar;
-        let query_response = load_file("test-data/forex/central-bank-of-myanmar.json");
-        let timestamp: u64 = 1656374400;
-        let extracted_rates = myanmar.extract_rate(&query_response, timestamp);
-        assert!(matches!(extracted_rates, Ok(ref rates) if rates["EUR"] == 1_059_297_297));
-        assert!(matches!(extracted_rates, Ok(ref rates) if rates["JPY"] == 7_369_729));
-    }
-
-    /// The function tests if the [CentralBankOfBosniaHerzegovina] struct returns the correct forex rate.
-    #[test]
-    fn extract_rate_from_bosnia() {
-        let bosnia = CentralBankOfBosniaHerzegovina;
-        let query_response = load_file("test-data/forex/central-bank-of-bosnia-herzegovina.json");
-        let timestamp: u64 = 1656374400;
-        let extracted_rates = bosnia.extract_rate(&query_response, timestamp);
-        assert!(matches!(extracted_rates, Ok(ref rates) if rates["EUR"] == 1_057_200_262));
-        assert!(matches!(extracted_rates, Ok(ref rates) if rates["JPY"] == 7_380_104));
-    }
-
-    /// The function tests if the [EuropeanCentralBank] struct returns the correct forex rate.
-    #[test]
-    fn extract_rate_from_ecb() {
-        let ecb = EuropeanCentralBank;
-        let query_response = load_file("test-data/forex/central-bank-of-europe.xml");
-        let timestamp: u64 = 1664755200;
-        let extracted_rates = ecb.extract_rate(&query_response, timestamp);
-
-        assert!(matches!(extracted_rates, Ok(rates) if rates["EUR"] == 976_400_000));
-    }
-
-    /// The function tests if the [BankOfCanada] struct returns the correct forex rate.
-    #[test]
-    fn extract_rate_from_canada() {
-        let canada = BankOfCanada;
-        let query_response = load_file("test-data/forex/bank-of-canada.json");
-        let timestamp: u64 = 1656374400;
-        let extracted_rates = canada.extract_rate(&query_response, timestamp);
-
-        assert!(matches!(extracted_rates, Ok(rates) if rates["EUR"] == 1_052_938_432));
-    }
-
-    /// The function tests if the [CentralBankOfUzbekistan] struct returns the correct forex rate.
-    #[test]
-    fn extract_rate_from_uzbekistan() {
-        let uzbekistan = CentralBankOfUzbekistan;
-        let query_response = load_file("test-data/forex/central-bank-of-uzbekistan.json");
-        let timestamp: u64 = 1656374400;
-        let extracted_rates = uzbekistan.extract_rate(&query_response, timestamp);
-
-        assert!(matches!(extracted_rates, Ok(rates) if rates["EUR"] == 1_056_900_158));
-    }
 
     /// Tests that the [OneDayRatesCollector] struct correctly collects rates and returns them.
     #[test]
@@ -2120,23 +1504,6 @@ mod test {
             store.get(queried_timestamp, queried_timestamp, "EUR", USD),
             Err(GetForexRateError::InvalidTimestamp(_queried_timestamp))
         ));
-    }
-
-    /// This function tests that the forex sources can report the max response bytes needed to make a successful HTTP outcall.
-    #[test]
-    fn forex_max_response_bytes() {
-        let forex = Forex::MonetaryAuthorityOfSingapore(MonetaryAuthorityOfSingapore);
-        assert_eq!(forex.max_response_bytes(), 3 * ONE_KIB);
-        let forex = Forex::CentralBankOfMyanmar(CentralBankOfMyanmar);
-        assert_eq!(forex.max_response_bytes(), 3 * ONE_KIB);
-        let forex = Forex::CentralBankOfBosniaHerzegovina(CentralBankOfBosniaHerzegovina);
-        assert_eq!(forex.max_response_bytes(), 30 * ONE_KIB);
-        let forex = Forex::EuropeanCentralBank(EuropeanCentralBank);
-        assert_eq!(forex.max_response_bytes(), 3 * ONE_KIB);
-        let forex = Forex::BankOfCanada(BankOfCanada);
-        assert_eq!(forex.max_response_bytes(), 10 * ONE_KIB);
-        let forex = Forex::CentralBankOfUzbekistan(CentralBankOfUzbekistan);
-        assert_eq!(forex.max_response_bytes(), 30 * ONE_KIB);
     }
 
     #[test]
